requirements.txt
# Core web app
flask==3.0.2
python-dotenv
gunicorn==21.2.0

# Data / file processing
pandas==2.2.2
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
reportlab==4.0.7

# AI / LLM
groq==0.4.2
langchain==0.1.0
langchain-groq==0.0.1

# Flask internals
werkzeug==3.0.1
